requirements.txt
flask==3.0.3
selenium==4.21.0
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
matplotlib==3.8.4
seaborn==0.13.2
requests==2.31.0
gunicorn
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
colorama==0.4.6
Flask==3.0.3
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
schedule
beautifulsoup4==4.12.3
gunicorn
waitress==2.1.2
torch==2.7.0
requests==2.31.0